+++ packages/react/src/App.jsx
@@ -3,9 +3,11 @@ import React, { useState, useEffect } from 'react';
 
 // Local  imports
 import DocumentData from './Types/DocumentData';
-import PageBody from './Components/PageBody';
+import PageBody from './Components/Molecules/PageBody';
 import { Provider } from './Components/Context';
 import { initOnlineStatus } from './helpers/electron';
+import './Components/Style/atoms.scss';
+import './Components/Style/global.scss';
 
 initOnlineStatus();
 const eventSourceData = {};
