+++ packages/react/src/Components/Molecules/index.js
@@ -1,0 +1,11 @@
+export { default as ImageCard } from './ImageCard';
+export { default as DocumentImage } from './DocumentImage';
+export { default as Header } from './Header';
+export { default as ImagePanel } from './ImagePanel';
+export { default as InfoTable } from './InfoTable';
+export { default as InfoTabs } from './InfoTabs';
+export { default as LiveImage } from './LiveImage';
+export { default as MatchValue } from './MatchValue';
+export { default as MRZPanel } from './MRZPanel';
+export { default as PageBody } from './PageBody';
+export { default as PhotoHeaders } from './PhotoHeaders';
